add schnorrsig feature
